backend/requirements.txt
torch>=2.2.0
diffusers>=0.26
transformers>=4.40
accelerate>=0.27
fastapi>=0.111
uvicorn[standard]>=0.29
pillow>=10.2
scipy>=1.12
python-multipart>=0.0.6
pydantic>=2.0.0
controlnet_aux>=0.0.6
numpy>=1.24.0
safetensors>=0.3.1 